Generate real Hyperledger chaincode (#122)
* Swap mock classes out for real Hyperledger chaincode library.

(It doesn't actually compile yet, but committing this for posterity.)

* Introduce intermediate class to handle chaincode interface.

(At least the generated compiles now! A bit of a different story getting
 it to run on hyperledger, but hey, progress.)

* (#107) Now can successfully run and be initialized on Fabric.

I had to add a lot of .jars to the manifest that are required by
Hyperledger, or else the jar wouldn't start because it didn't know where
to find the required classes.

* Start making gradle script to handle dependencies.

* Fix 'can't find main class' issue by creating a fat JAR.

* Create some shell scripts to ease testing.

Note: the 'startfabric.sh' one won't work multiple times unless you run
it with the 'clean' option, which just deletes your whole hyperledger
folder and recreates it. This is because hyperledger doesn't let us
delete channels or chaincodes (which makes sense, given the nature of
blockchain), but we want to test stuff without saving all our files.

* Add compiler flag to toggle mock/hyperledger chaincode.

* Remove generateInvokeMethod -- it was never used

Instead, invoke() in the base class delegates to the already-generated
run() method. I don't know why I added this other method.

* Make hyperledger flag variable less confusing.

* rename 'bleh' file to something more useful

* Hopefully make sed commands in script more portable.

Apparently GNU sed and OSX sed differ in their interpretation of "\+" in
non-extended mode.

* actually just use * for the regexes.

* run gradle script automatically when compiling

Now the output JAR when running in Hyperledger will always be located at
`buildscript/build/libs/chaincode.jar`.

* add usage to invoke.sh

* move Gradle output; change where it reads source files from

* Add instantiate.sh to make it easier

* make paths relative to startfabric.sh script

* print 'true' instead of 'false' when doing Hyperledger chaincode

* make invokeConstructor protected for real chaincode as well

* make startfabric.sh take path to chaincode.jar as argument

Fixes #107 .