Issue 157: ReadOnlyState support (#160)
Added support for "this" arguments that have no permission specified, in which case ReadOnlyState is inferred. "this" arguments that are missing are still treated as Shared.

This change includes a few small improvements that I made along the way, including removal of the availableIn field from transactions and a fix to the parser to no longer use "THIS" as a contract name.